pyproject.toml
[project]
name = "django-tws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "alex", email = "alex@pdp7.net" }
]
requires-python = ">=3.11"
dependencies = [
    "django-environ>=0.12.0",
]

[project.scripts]
django-tws = "django_tws:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
